# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/comprobantes": {
      "post": {
        "operationId": "ComprobantesController_crear",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComprobanteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Crear un comprobante en BORRADOR con sus líneas. No valida partida doble — ese chequeo corre al contabilizar.",
        "tags": [
          "Asientos contables"
        ]
      },
      "get": {
        "operationId": "ComprobantesController_listar",
        "parameters": [
          {
            "name": "periodoFiscalId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipo",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "APERTURA",
                "DIARIO",
                "INGRESO",
                "EGRESO",
                "AJUSTE",
                "TRASPASO",
                "CIERRE"
              ]
            }
          },
          {
            "name": "estado",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "BORRADOR",
                "CONTABILIZADO",
                "BLOQUEADO"
              ]
            }
          },
          {
            "name": "fechaDesde",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-04-01",
              "type": "string"
            }
          },
          {
            "name": "fechaHasta",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-04-30",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Busca en numero y glosa (case-insensitive)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "incluirAnulados",
            "required": false,
            "in": "query",
            "description": "Incluir comprobantes anulados en el listado (default: false)",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListarComprobantesResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Listar comprobantes del tenant con paginación y filtros (periodoFiscalId, tipo, estado, rango de fechas, texto libre).",
        "tags": [
          "Asientos contables"
        ]
      }
    },
    "/api/comprobantes/{id}": {
      "get": {
        "operationId": "ComprobantesController_obtener",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Detalle completo del comprobante con sus líneas.",
        "tags": [
          "Asientos contables"
        ]
      },
      "patch": {
        "operationId": "ComprobantesController_actualizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarContabilizadoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Actualizar un comprobante. Si está en BORRADOR, actualiza parcialmente (todos los campos opcionales). Si está CONTABILIZADO y el período está abierto, edita cabecera y/o líneas — requiere permiso adicional `contabilidad.asientos.edit-posted` (§4.3 CLAUDE.md). El número correlativo es INMUTABLE. Rechaza BLOQUEADO o anulados.",
        "tags": [
          "Asientos contables"
        ]
      },
      "delete": {
        "operationId": "ComprobantesController_eliminar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Eliminar físicamente un BORRADOR. Un CONTABILIZADO/BLOQUEADO no se borra: se anula con /anular.",
        "tags": [
          "Asientos contables"
        ]
      }
    },
    "/api/comprobantes/{id}/contabilizar": {
      "post": {
        "operationId": "ComprobantesController_contabilizar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Transicionar BORRADOR → CONTABILIZADO. Valida partida doble, asigna número atómico {prefijo}{YY}{MM}-{correlativo:6} y registra auditoría.",
        "tags": [
          "Asientos contables"
        ]
      }
    },
    "/api/comprobantes/{id}/anular": {
      "post": {
        "operationId": "ComprobantesController_anular",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnularComprobanteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Anular un CONTABILIZADO con flag anulado=true (CLAUDE.md §4.7). No genera contra-asiento. El número del original se preserva.",
        "tags": [
          "Asientos contables"
        ]
      }
    },
    "/api/comprobantes/{id}/auditoria": {
      "get": {
        "operationId": "ComprobantesController_obtenerAuditoria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Historial de auditoría del comprobante: cada acción con usuario, timestamp y diff.",
        "tags": [
          "Asientos contables"
        ]
      }
    },
    "/api/comprobantes/{comprobanteId}/documentos-fisicos": {
      "get": {
        "operationId": "ComprobantesController_listarDocumentosFisicos",
        "parameters": [
          {
            "name": "comprobanteId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Listar los documentos físicos asociados al comprobante (REQ-A-09).",
        "tags": [
          "Asientos contables"
        ]
      },
      "post": {
        "operationId": "ComprobantesController_asociarDocumentosFisicos",
        "parameters": [
          {
            "name": "comprobanteId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AsociarDocumentosDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Asociar uno o más documentos físicos a un comprobante. Aditiva e idempotente (REQ-A-01). En BORRADOR no exige permisos extra. En CONTABILIZADO con período abierto edita la asociación (§4.3 CLAUDE.md) — requiere permiso adicional `contabilidad.asientos.edit-posted`. Rechaza período cerrado/bloqueado y comprobantes anulados.",
        "tags": [
          "Asientos contables"
        ]
      }
    },
    "/api/comprobantes/{comprobanteId}/documentos-fisicos/{documentoFisicoId}": {
      "delete": {
        "operationId": "ComprobantesController_desasociarDocumentoFisico",
        "parameters": [
          {
            "name": "comprobanteId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "documentoFisicoId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Desasociar un documento físico de un comprobante. En BORRADOR no exige permisos extra. En CONTABILIZADO con período abierto (§4.3 CLAUDE.md) requiere `contabilidad.asientos.edit-posted`. Rechaza período cerrado/bloqueado y comprobantes anulados (REQ-A-02/03).",
        "tags": [
          "Asientos contables"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-Tenant SaaS API",
    "description": "Production-ready NestJS API with multi-tenancy, authentication, RBAC, billing, and more.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      },
      "X-Tenant-ID": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID",
        "description": "Tenant ID for multi-tenant operations"
      }
    },
    "schemas": {
...
      "CreateLineaDto": {
        "type": "object",
        "properties": {
          "cuentaId": {
            "type": "string",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "contactoId": {
            "type": "string",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "description": "Requerido si cuenta.requiereContacto al contabilizar"
          },
          "moneda": {
            "type": "string",
            "enum": [
              "BOB",
              "USD"
            ],
            "example": "BOB"
          },
          "debito": {
            "type": "string",
            "example": "1000.00",
            "description": "Monto DEBE en moneda original (string)"
          },
          "credito": {
            "type": "string",
            "example": "0",
            "description": "Monto HABER en moneda original (string)"
          },
          "tipoCambio": {
            "type": "string",
            "example": "1",
            "description": "Tipo de cambio a BOB (1 si moneda=BOB)"
          },
          "debitoBob": {
            "type": "string",
            "example": "1000.00",
            "description": "Monto DEBE convertido a BOB"
          },
          "creditoBob": {
            "type": "string",
            "example": "0",
            "description": "Monto HABER convertido a BOB"
          },
          "glosaLinea": {
            "type": "string",
            "description": "Glosa específica de la línea (opcional)"
          }
        },
        "required": [
          "cuentaId",
          "moneda",
          "debito",
          "credito",
          "tipoCambio",
          "debitoBob",
          "creditoBob"
        ]
      },
      "CreateComprobanteDto": {
        "type": "object",
        "properties": {
          "tipo": {
            "type": "string",
            "enum": [
              "APERTURA",
              "DIARIO",
              "INGRESO",
              "EGRESO",
              "AJUSTE",
              "TRASPASO",
              "CIERRE"
            ],
            "example": "DIARIO"
          },
          "fechaContable": {
            "type": "string",
            "example": "2026-04-22",
            "description": "Fecha contable en formato ISO YYYY-MM-DD"
          },
          "glosa": {
            "type": "string",
            "example": "Venta al contado a cliente X"
          },
          "monedaPrincipal": {
            "type": "string",
            "enum": [
              "BOB"
            ],
            "default": "BOB"
          },
          "tipoCambioReexpresion": {
            "type": "string",
            "example": "6.96",
            "description": "T/C de presentación (re-expresión). Decimal estrictamente positivo. Omitir para usar default 1."
          },
          "lineas": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateLineaDto"
            }
          }
        },
        "required": [
          "tipo",
          "fechaContable",
          "glosa",
          "lineas"
        ]
      },
      "ContactoResumenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string",
            "example": "Avícola Sur S.R.L."
          }
        },
        "required": [
          "id",
          "nombre"
        ]
      },
      "DocumentoRespaldoResumenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tipoNombre": {
            "type": "string",
            "example": "Factura"
          },
          "numero": {
            "type": "string",
            "example": "0042"
          }
        },
        "required": [
          "id",
          "tipoNombre",
          "numero"
        ]
      },
      "ComprobanteListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "APERTURA",
              "DIARIO",
              "INGRESO",
              "EGRESO",
              "AJUSTE",
              "TRASPASO",
              "CIERRE"
            ]
          },
          "numero": {
            "type": "string",
            "nullable": true,
            "example": "I2604-000042"
          },
          "estado": {
            "type": "string",
            "enum": [
              "BORRADOR",
              "CONTABILIZADO",
              "BLOQUEADO"
            ]
          },
          "fechaContable": {
            "type": "string",
            "example": "2026-04-22"
          },
          "periodoFiscalId": {
            "type": "string"
          },
          "glosa": {
            "type": "string"
          },
          "monedaPrincipal": {
            "type": "string",
            "enum": [
              "BOB",
              "USD"
            ]
          },
          "tipoCambioReexpresion": {
            "type": "string",
            "example": "1.00000000"
          },
          "totalDebitoBob": {
            "type": "string",
            "example": "1000.00"
          },
          "totalCreditoBob": {
            "type": "string",
            "example": "1000.00"
          },
          "anulado": {
            "type": "boolean"
          },
          "fechaAnulacion": {
            "type": "string",
            "nullable": true
          },
          "anuladoPorUserId": {
            "type": "string",
            "nullable": true
          },
          "motivoAnulacion": {
            "type": "string",
            "nullable": true
          },
          "createdByUserId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "contactos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactoResumenDto"
            }
          },
          "documentosRespaldo": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentoRespaldoResumenDto"
            }
          }
        },
        "required": [
          "id",
          "tipo",
          "numero",
          "estado",
          "fechaContable",
          "periodoFiscalId",
          "glosa",
          "monedaPrincipal",
          "tipoCambioReexpresion",
          "totalDebitoBob",
          "totalCreditoBob",
          "anulado",
          "fechaAnulacion",
          "anuladoPorUserId",
          "motivoAnulacion",
          "createdByUserId",
          "createdAt",
          "updatedAt",
          "contactos",
          "documentosRespaldo"
        ]
      },
      "ListarComprobantesResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComprobanteListItemDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit"
        ]
      },
      "EditarContabilizadoDto": {
        "type": "object",
        "properties": {
          "tipo": {
            "type": "string",
            "enum": [
              "APERTURA",
              "DIARIO",
              "INGRESO",
              "EGRESO",
              "AJUSTE",
              "TRASPASO",
              "CIERRE"
            ],
            "example": "DIARIO"
          },
          "fechaContable": {
            "type": "string",
            "example": "2026-04-22",
            "description": "Fecha contable en formato ISO YYYY-MM-DD"
          },
          "glosa": {
            "type": "string",
            "example": "Venta al contado a cliente X"
          },
          "monedaPrincipal": {
            "type": "string",
            "enum": [
              "BOB"
            ],
            "default": "BOB"
          },
          "tipoCambioReexpresion": {
            "type": "string",
            "example": "6.96",
            "description": "T/C de presentación (re-expresión). Decimal estrictamente positivo. Omitir para usar default 1."
          },
          "lineas": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateLineaDto"
            }
          },
          "numero": {
            "type": "string",
            "description": "Número correlativo. INMUTABLE — si se envía distinto al actual, la petición será rechazada con 409 (§4.9 CLAUDE.md). Omitir en producción."
          },
          "motivo": {
            "type": "string",
            "description": "Motivo del cambio, registrado en comprobantes_audit. Si se omite, el trigger registrará motivo NULL (auditoría silenciosa). Mínimo 3 caracteres si se envía.",
            "example": "Corrección de glosa — error tipográfico en el nombre del cliente",
            "minLength": 3,
            "maxLength": 500
          }
        }
      },
      "AnularComprobanteDto": {
        "type": "object",
        "properties": {
          "motivo": {
            "type": "string",
            "example": "Error en la imputación al cliente",
            "description": "Motivo de la anulación, mínimo 10 caracteres (visible en auditoría)",
            "minLength": 10,
            "maxLength": 500
          }
        },
        "required": [
          "motivo"
        ]
      },
      "AsociarDocumentosDto": {
        "type": "object",
        "properties": {
          "documentoFisicoIds": {
            "description": "IDs de los documentos físicos a asociar al comprobante. La operación es aditiva e idempotente.",
            "example": [
              "7c0e5b2a-2f3d-4d6e-9a1b-1c2d3e4f5a6b"
            ],
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "documentoFisicoIds"
        ]
      },
...
    }
  }
}